Fix for bug 605283: SDK tutorial should explain about content scripts
Includes a new Markdown file "web-content.md" which talks about how
to use content scripts. The text also uses the "Reddit" example which
is removed from the panel.md documentation.

This change also updates "index.html" to refer to the new tutorial,
section, and "programs.md" the preceding tutorial section, to refer
to the new tutorial.

 Committer: Will <wbamberg@mozilla.com>

 On branch bug605283
 Changes to be committed:
   (use "git reset HEAD <file>..." to unstage)

	modified:   packages/addon-kit/docs/panel.md
	modified:   static-files/index.html
	modified:   static-files/md/dev-guide/programs.md
	new file:   static-files/md/dev-guide/web-content.md